requirements.txt
connectomics
dataclasses-json>=0.5.6
absl-py>=0.9.0
jax>=0.2.25
numpy>=1.21.5
opencv-python>=4.5.5.62
scipy>=1.2.3
scikit-image>=0.17.2
